Extract generatorform to own component + fix textAlign property of header
